+++ CHANGELOG.md
@@ -4,8 +4,9 @@ Continuuity Loom CHANGELOG
 Unreleased
 ----------
 - Settable max for clusters and nodes in a tenant ( Issues: #302 )
+- Queue Metrics ( Issues: #301 )
 - Provisioner Multi Tenancy ( Issues: #290, #296 )
-- Multi Tenancy ( Issues: #239, #253, #277, #284 )
+- Multi Tenancy ( Issues: #239, #253, #277, #284, #292 )
 - Ruby testing with rspec/rubocop ( Issues: #128 #132 #133 )
 - Nginx support for clusters ( Issues: #131 )
 - Sensu monitoring support ( Issues: #167 )
@@ -13,8 +14,11 @@ Unreleased
 - Ability to sync an active cluster's template to its current version ( Issues: #214 )
 - Cluster owner macro ( Issues: #221 )
 - Loom Firewall updates ( Issues: #226 #227 )
+- Multi Tenancy ( Issues: #239 #253 #277 #284 )
 - Include automator and provider types in export and import ( Issues: #246 )
 - Upgrade script ( Issues: #258 )
+- Provisioner Multi Tenancy ( Issues: #290 #296 )
+- Updated provider plugin using fog ( Issues: #288 #310 )
 
 v0.9.7 (May 27, 2014)
 ---------------------